+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 23.8.2
 - Fixed rating feedback widget event key for widget closed event
+- Added `testingDownloadExperimentInformation:` in remote config interface
+- Added `testingGetAllExperimentInfo:` in remote config interface
 
 ## 23.8.1
 - Expanded feedback widget functionality. Added ability to use rating widgets.